Add UDP outgoing test with separate pods.  (#445)
* UDP outgoing test with separate pods.

Set up two services, one of them only reachable from within the cluster.
That internal service runs a server that logs UDP messages.
Impersonate the other service with mirrord, running an application that
sends a UDP message to the internal one. Read the logs from the
internal service to verify the outgoing UDP message arrived.

* Rename EchoService into KubeService.

It's not always an http-echo service anymore, so change to a more
generic name.