Add layer integration test.
Add a feature for the layer to connect over tcp instead of starting
an agent with kuberenetes.

Load the layer into a flask server process, have it connect to the
tests code over TCP, verify correct messages from layer, send
(fake) TCP notifications to the layer and check app output.

Since the `build_mirrord` job builds the library anyways, use that
artifact for running the test on ubuntu. For MacOS just build the
library in a step in the test job.

Make rebuilding the library from code the defualt when running the
integration tests, with an the option to use a prebuilt lib, for
CI. Because building locally is fast and that way when testing
locally you always test the latest layer code.

Make the `test_mirrord_layer` job only run unit tests.

Add CI dummies for the new jobs.

Co-authored-by: Aviram Hassan <aviram@metalbear.co>