Bypass SIP on MacOS (#706)
Part of https://github.com/metalbear-co/mirrord/issues/649.

On MacOS, test if the file that the user specified to run is protected by SIP, or is a script with a shebang that leads to such a binary (also via multiple shebang scripts). If it is, create an unprotected version of the binary (and the shebang scripts that lead to it) and run that instead.

This does not yet bypass SIP on binaries that the executed application executes.

Co-authored-by: t4lz <t4lz.git@gmail.com>