Intercept argv and remove mirrord's temp dir (#759)
When an executable script `my_script` with a shebang `#!/usr/bin/env some_interpreter` is executed, it resolves to 
```bash
/usr/bin/env some_interpreter my_script
```
So currently if we patch it and save a patch version in `mirrord-tmp/`, running `/mirrord-tmp/my_script` resolves to 
```bash
/usr/bin/env some_interpreter /mirrord-temp/my_script
```
so what then happens when we run `my_script` with the mirrord CLI (if `some_interpreter` is a binary executable that does not have SIP):
1. We read `my_script` and its shebang.
2. We create a version of `my_script` in `/mirrrord-tmp/path/to/my_script` with the shebang `#!/mirrord-tmp/usr/bin/env some_interpreter`, and create a SIP-free version of `env` in `/mirrord-tmp/usr/bin/env`.
3. We execute `/mirrrord-tmp/path/to/my_script` which resolves to `/mirrord-tmp/usr/bin/env my_interpreter /mirrrord-tmp/path/to/my_script`
4. patched `env` is executed with loaded layer.
5. `env` calls `execve` with path `my_interpreter` and argv: `["my_interpreter", "/mirrrord-tmp/path/to/my_script"]`.

This is a problem when running `node` as the interpreter (`#!/usr/bin/env node`), since node determines the location of modules relatively to the executed file, which it receives as an argument (here argv[1]). In our example node would search for modules relative to `/mirrrord-tmp/path/to/my_script`, instead of relative to `/path/to/my_script`.
This PR suggests a mitigation that problem: 
In the `execve` hook: go over the arguments in `argv` and strip the path to our temp directory from them if it's there.

That way node reads the original script, not the patched one, and can find the required modules.