Send a `RemoteResult` in the `Subscribed` message.
Send back the subscribed port in a `RemoteResult` from agent to
layer. If the layer gets an error on the subscription message,
it raises an error which stops the layer. The stealer returns an
error if it receives a subscription request to a port that is
already being stolen by another client (another layer).