Start e2e tests in parallel to integration tests.
Currently the e2e tests depend on the integration tests, so that
they are only run if integration tests are successful.
Sometimes (but not always) the integration tests finish after the
other dependencies of the e2e tests, so by removing the dependency
the e2e tests would start earlier in those cases and the whole
github action would end earlier.