Move the stream-peeking into the filter task.
Also simplify the filter task.

Since creating a reversible stream can wait for up to 10 seconds,
move that into the separate filter task.
This also has the effect of not the stealer not being able to
fail and stop because of an error in dealing with one HTTP
connection.

And since there's no more any logic in the HttpFilterBuilder,
just convert it to a function.