Hook `posix_spawn` with same logic as `execve`. (#988)
Fixes https://github.com/metalbear-co/mirrord/issues/987.

The problem was that this newer node version does not use libc's `execve` to spawn a new process but libuv's `uv_spawn`, which calls `posix_spawn`. This means that our `execve` was not being called and so we could not patch SIP binaries, and we would therefore not load into the new process.
So now we also hook `posix_spawn`.

This PR also adds a regression test. The test was verified to be meaningful by checking that it indeed fails without the new hook: https://github.com/t4lz/mirrord/actions/runs/4028503663/jobs/6925471767#step:19:283.