Add agent-test-error feature.
Add a configuration that enables to make the agent return an error
after accepting the first client.
Only include that option in debug builds, except for in the agent,
since it's usually tested with a release build, so it's also
included in release builds. However users cannot use this feature
on released versions because it's not present in the cli or in the
configuration.