Unpause on Agent Error (#1116)
* unpause the target container on exit if paused.

* Add agent-test-error feature.

Add a configuration that enables to make the agent return an error
after accepting the first client.
Only include that option in debug builds, except for in the agent,
since it's usually tested with a release build, so it's also
included in release builds. However users cannot use this feature
on released versions because it's not present in the cli or in the
configuration.

* enable putting test k8s resources in custom namespaces.

* unpause e2e test.

* CR: agent cli docs.

Co-authored-by: meowjesty <43983236+meowjesty@users.noreply.github.com>

* CR: docs linebreaks

Co-authored-by: meowjesty <43983236+meowjesty@users.noreply.github.com>

* CR: docs.

---------

Co-authored-by: meowjesty <43983236+meowjesty@users.noreply.github.com>