Allow x86_64 and arm64 when patching a thin binary (#2187)
In a thin binay there is no arch choice anyway, so patch if we know
how to patch that binary type (we know how to patch `x86_64` and
`arm_64`).

Up until now we would try to (imperfectly) forbid patching thin
`arm64` binaries on `x86_64` systems, but there is no point in
that, as that is not a valid situation anyways, so it should not
happen, and if it does happen, that's not a mirrord error, so we
can let it fail naturally as it would without mirrord.