Add new features enum and field for operator
The old enum cannot be extended as it would break old clients.
So we define a new extensible enum and a new field for it.
We make all of those fields private so that users of that struct
only access them via a getter that handles objects deserialized
from either old or new versions.