feature(jsonschema): `target` Simple Case is 'string'
fixes #2188

For JSONSchema used by VSCode, configure schemars for the simple `target` case to allow just a string.
The complex/Advanced case is still going to be the object, although that is still broken for `target.path` being a string.

Currently there is no validation of what the string is. I could probably hardcode a regex for container/pod/deployment/rollout,
but that seems like it could get out of sync in the future. Unsure if there's a way to just coerce Target variants to strings
and regex on that, but that seems like it could be complicated.

## Examples

This is now valid and doesn't throw warnings in vscode:

```json
{
	"target": "deploy/my-deploy"
}
```

This is still invalid / throws warnings per the schema:

```json
{
	"target": {
		  "path": "deploy/my-deploy"
	}
}
```