Simplify local-read regexes and anchor current dir and exe (#2921)
* Changed `^.+` to just `.`

Since we're searching for matches inside a given string, and not
trying to match the whole string, we don't need to anchor the regex
to the beginning of the line in this case.

* Remove `^.*` from regex start

We don't need to require a line start followed by any or no characters.

* Remove `.*$` from regex end

there will always be any or no character (`.*`) and then a line-end (`$`)
after any pattern, so we don't need that part in the regex.

* Anchor current dir to start, current exe to end.

Instead of ignoring any path that contains them anywhere.

* CR: match period

Co-authored-by: Michał Smolarek <34063647+Razz4780@users.noreply.github.com>

* CR: stricter current_dir matching

Require a '/' or '$' after the current dir's path.
Handling both cases - when current_dir returns a trailing / and when
not.

---------

Co-authored-by: Michał Smolarek <34063647+Razz4780@users.noreply.github.com>