Create ADOPTERS.md and CTA to README.md (#3696)
* Create ADOPTERS.md to showcase users of mirrord

Added a list of organizations using mirrord, based on 'user stories' on MetalBear.com, along with their use cases and links to case studies on MetalBear.com

* Enhance README with user feedback and adoption info

Added CTA on README, to prompt adopters to share their usage of mirrord publicly.

* Revise user feedback section in README

Refined the wording of the lede into the CTA.

* Revise ADOPTERS.md to update links and format

Removed Use Case column to avoid confusion and repetition

* Clarify 'More Info' column in ADOPTERS.md

Updated the ADOPTERS.md file to clarify the 'More Info' column as optional and added a note about it.

* Improve wording in README about ADOPTERS.md

Reword the section about ADOPTERS.md for clarity.

* Add changelog

---------

Co-authored-by: Eyal Bukchin <eyal@metalbear.co>